VIX-2715 - Add an Effect Preview Toolstrip and Effect Preview button and remove all Effect Preview code that was used in the Effect editor. I left the ToggleSwitch code that was used in the Effect editor as I thought it may be good to leave just in case we can use this control later. Hope I got everything.
