+++ Common/Resources/Properties/Resources.resx
@@ -1108,4 +1108,7 @@
   <data name="zoom_time_out" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\zoom_time_out.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
+  <data name="Binoculars" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\Binoculars.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
 </root>
