+++ Common/Resources/Resources.csproj
@@ -380,6 +380,7 @@
     <None Include="Resources\TimingSpeed_350.png" />
     <None Include="Resources\TimingSpeed_250.png" />
     <None Include="Resources\TimingSpeed.png" />
+    <None Include="Resources\Binoculars.png" />
     <Content Include="table_chart_left.png" />
     <Resource Include="folder_open.png" />
     <None Include="hourglass_link.png" />
